switch to updated TorchFCPE + dropout1d support
- Removal of applio's fcpe and fumiama rvc's fcpe ( " fumi_fcpe " ).
- TorchFCPE becomes the main and only fcpe.
- Filter radius now has an actual use. Check the note.
- Added support for Dropout1d in nsf hifigan generator

NOTE:
- TorchFCPE's current version is 0.0.4 however since the last release, some changes / updates were done ( on official repository ).
Those changes are present here, in this fork's fcpe.
- Filter Radius is now controlling the radius / 'threshold' for FCPE.
More info on that in description you can read when you open the inference tab in fork. ( ps. it's only for fcpe. )
- nsf hifigan now uses Dropout1d of which rate you can ( for now ) control through setting the 'dropout_value' global from within the training script.

I did some prior testing and stuff works but as always, in case of any issues, lemme know asap. Cheers.

+++ rvc/infer/infer.py
@@ -206,7 +206,7 @@ class VoiceConverter:
         split_audio: bool = False,
         f0_autotune: bool = False,
         f0_autotune_strength: float = 1,
-        filter_radius: int = 3,
+        filter_radius: float = 3.0,
         embedder_model: str = "contentvec",
         embedder_model_custom: str = None,
         clean_audio: bool = False,
@@ -222,7 +222,7 @@ class VoiceConverter:
 
         Args:
             pitch (int): Key for F0 up-sampling.
-            filter_radius (int): Radius for filtering.
+            filter_radius (float): Radius for filtering.
             index_rate (float): Rate for index matching.
             volume_envelope (int): RMS mix rate.
             protect (float): Protection rate for certain audio segments.
